Clean up code and add room util file
Clean up the code, especially the main.js. Also added a few convenience functions.

Moved the design function to the spawner.

Also remove the hard reference to the room in the main loop, so now the code should work with multiple rooms (although there is no movement between rooms yet)