Slight strategic changes to harvest more
There were only two harvesters which was subobtimal. Upped it to 5.

Also, the harvesters will now move to the spawner and not block the resource if they are idlingg because all dropoffs are full. Added a room.find_spawner function that will get the spawner they will move to.

also, Builders and harvesters carry/move more and work less, which seems better because they were soooo slow